Function definitions changed and variable issues of transfer from host to device fixed

+++ interior_flux_dGxneg_mod.h
@@ -140,7 +140,7 @@ void interior_dGx_neg(double *G, int i)
 }
 //
 
-__global__ void interior_dGx_neg_cuda(points &point, int power, double VL_CONST,double pi,int interior_points,int *interior_points_index)
+__global__ void interior_dGx_neg_cuda(points &point, double power, double VL_CONST,double pi,int interior_points,int *interior_points_index)
 //
 //
 {
