+++ outer_flux_dGxpos_mod.h
@@ -142,7 +142,7 @@ void outer_dGx_pos(double *G, int i)
 
 
 
-__global__ void outer_dGx_pos_cuda(points &point,int power, double VL_CONST,double pi,int outer_points,int *outer_points_index)
+__global__ void outer_dGx_pos_cuda(points &point,double power, double VL_CONST,double pi,int outer_points,int *outer_points_index)
 //
 //
 {
