+++ outer_flux_dGypos_mod.h
@@ -144,7 +144,7 @@ void outer_dGy_pos(double *G, int i)
 
 
 
-__global__ void outer_dGy_pos_cuda(points &point,int power, double VL_CONST,double pi,int outer_points,int *outer_points_index)
+__global__ void outer_dGy_pos_cuda(points &point,double power, double VL_CONST,double pi,int outer_points,int *outer_points_index)
 //
 //
 {
