+++ wall_flux_dGypos_mod.h
@@ -142,7 +142,7 @@ void wall_dGy_pos(double *G, int i)
 }
 //
 
-__global__ void wall_dGy_pos_cuda(points &point,int power, double VL_CONST,double pi,int wall_points,int *wall_points_index)
+__global__ void wall_dGy_pos_cuda(points &point,double power, double VL_CONST,double pi,int wall_points,int *wall_points_index)
 //
 //
 {
