+++ wall_flux_dGzneg_mod.h
@@ -140,7 +140,7 @@ void wall_dGz_neg(double *G, int i)
 }
 //
 
-__global__ void wall_dGz_neg_cuda(points &point, int power, double VL_CONST,double pi,int wall_points,int *wall_points_index)
+__global__ void wall_dGz_neg_cuda(points &point, double power, double VL_CONST,double pi,int wall_points,int *wall_points_index)
 //
 {
 	int ind = blockIdx.x * blockDim.x + threadIdx.x;
